path: Fix an integer overflow in l_path_get_mtime
time_t may be 32 bits wide (also signed) so we need to cast
timespec.tv_sec to uint64_t before multiplying by 1000000.

+++ ell/path.c
@@ -162,7 +162,8 @@ LIB_EXPORT uint64_t l_path_get_mtime(const char *path)
 	if (ret < 0)
 		return L_TIME_INVALID;
 
-	return sb.st_mtim.tv_sec * 1000000 + sb.st_mtim.tv_nsec / 1000;
+	return (uint64_t) sb.st_mtim.tv_sec * 1000000 +
+		sb.st_mtim.tv_nsec / 1000;
 }
 
 /**
